+++ pontis-ui/src/components/Asset/AssetLogo.tsx
@@ -1,19 +1,6 @@
 import React, { useState } from "react";
 import { AssetKeyT } from "../../hooks/oracle";
-
-/**
- * Returns name of the logo's image.
- * @param {AssetKeyT} assetKey string containing the coin abbreviation delimited by a '/'
- * @return {string} returns string that matches corresponding file name in /assets/logos
- */
-function getLogoPath(assetKey: AssetKeyT): string {
-  const indexOfSlash = assetKey.indexOf("/");
-  const logo =
-    indexOfSlash > 0
-      ? assetKey.slice(0, indexOfSlash).toLowerCase()
-      : "fallback";
-  return `/assets/currencies/${logo}.svg`;
-}
+import { getLogoPath } from "../../../utils/mappings";
 
 interface AssetLogoProps {
   assetKey: AssetKeyT;
