+++ pontis-ui/src/components/Details/DetailCard.tsx
@@ -30,16 +30,9 @@ const DetailCard: React.FC<DetailCardProps> = ({
         </p>
       ) : (
         <div className="flex flex-grow flex-row-reverse items-end justify-between sm:flex-col sm:items-start">
-          {/* <div className="flex flex-row items-center justify-between"> */}
           <div className="rounded-lg bg-slate-300 bg-opacity-30 p-3 group-hover:bg-slate-50">
             {img}
           </div>
-          {/* This is an indicator for changing prices */}
-          {/* <div className="inline-flex text-sm text-gray-600 group-hover:text-gray-200 sm:text-base">
-              <MinusCircleIcon className="mr-2 h-6 w-6 stroke-green-500 group-hover:stroke-slate-200" />
-              12%
-            </div> */}
-          {/* </div> */}
           <div className="font-mono text-3xl font-bold text-slate-700 group-hover:text-slate-50 sm:mt-10 sm:text-4xl md:mt-12 xl:text-5xl">
             {toDisplay}
           </div>
