+++ pontis-ui/src/components/Navigation/NavFooter.tsx
@@ -19,7 +19,7 @@ const content: FooterColumn[] = [
     links: [
       {
         title: "Documentation",
-        href: "https://docs.empiric.network/",
+        href: "https://docs.empiric.network/quickstart",
         external: true,
       },
       {
