+++ pontis-ui/src/pages/about.tsx
@@ -32,7 +32,7 @@ const AboutPage = () => {
             icon: ChatIcon,
           }}
           secondaryAction={{
-            href: "https://docs.empiric.network/",
+            href: "https://docs.empiric.network/quickstart",
             actionText: "Read the docs",
             icon: CodeIcon,
           }}
