+++ pontis-ui/src/pages/features.tsx
@@ -15,7 +15,7 @@ const FeaturesPage = () => {
           title="Reimagining oracles"
           subtitle="Meet our compute engine"
           text="Our features empower native protocols to realize their ambitious potential."
-          href="https://docs.empiric.network/"
+          href="https://docs.empiric.network/quickstart"
           hrefText="Integrate verifyable data into your project"
         />
         <FeaturesDisplay />
@@ -42,7 +42,7 @@ const FeaturesPage = () => {
           title="Ready to get the data you need?"
           description="Leverage recent breakthroughs in zero knowledge computation by using verifyable and composable data in your application."
           mainAction={{
-            href: "https://docs.empiric.network/",
+            href: "https://docs.empiric.network/quickstart",
             actionText: "Read the docs",
             icon: CodeIcon,
           }}
